1、3.2.9.9完成_加油模块数据改完

+++ CttxApp/src/main/java/com/zantong/mobilecttx/fahrschule_v/SubjectOrderFragment.java
@@ -134,7 +134,7 @@ public class SubjectOrderFragment extends JxBaseRefreshFragment
             mTvPhone.setText(event.getPhone());
         }
         String price = event.getPrice();
-        mTvPrice.setText(FormatUtils.submitPrice(Float.valueOf(price)));
+        mTvPrice.setText(FormatUtils.submitPrice(price));
     }
 
     private String displayPriceValue(double discount, int couponType, int price) {
